#184 Move the language selection screen file to a better location
